+++ src/main/java/com/nemonotfound/nemos/mossy/vertical/slabs/datagen/ModRecipeProvider.java
@@ -1,6 +1,5 @@
 package com.nemonotfound.nemos.mossy.vertical.slabs.datagen;
 
-import com.nemonotfound.nemosmossyblocks.tag.ModItemTags;
 import com.nemonotfound.nemos.mossy.vertical.slabs.item.ModItems;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricRecipeProvider;
@@ -14,7 +13,9 @@ import org.jetbrains.annotations.NotNull;
 
 import java.util.concurrent.CompletableFuture;
 
-import static com.nemonotfound.nemosmossyblocks.item.ModItems.*;
+import static com.nemonotfound.nemos.mossy.blocks.item.Items.*;
+import static com.nemonotfound.nemos.mossy.blocks.registry.tag.ItemTags.MOSSY_PLANKS;
+
 
 public class ModRecipeProvider extends FabricRecipeProvider {
 
@@ -84,6 +85,64 @@ public class ModRecipeProvider extends FabricRecipeProvider {
                 this.stonecutterResultFromBase(RecipeCategory.BUILDING_BLOCKS, ModItems.PALE_MOSSY_TUFF_VERTICAL_SLAB, PALE_MOSSY_TUFF, 2);
                 this.stonecutterResultFromBase(RecipeCategory.BUILDING_BLOCKS, ModItems.PALE_MOSSY_TUFF_BRICK_VERTICAL_SLAB, PALE_MOSSY_TUFF_BRICKS, 2);
                 this.stonecutterResultFromBase(RecipeCategory.BUILDING_BLOCKS, ModItems.PALE_MOSSY_BRICK_VERTICAL_SLAB, PALE_MOSSY_BRICKS, 2);
+
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_OAK_VERTICAL_SLAB, CRIMSON_MOSSY_OAK_PLANKS);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_SPRUCE_VERTICAL_SLAB, CRIMSON_MOSSY_SPRUCE_PLANKS);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_BIRCH_VERTICAL_SLAB, CRIMSON_MOSSY_BIRCH_PLANKS);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_JUNGLE_VERTICAL_SLAB, CRIMSON_MOSSY_JUNGLE_PLANKS);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_ACACIA_VERTICAL_SLAB, CRIMSON_MOSSY_ACACIA_PLANKS);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_DARK_OAK_VERTICAL_SLAB, CRIMSON_MOSSY_DARK_OAK_PLANKS);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_MANGROVE_VERTICAL_SLAB, CRIMSON_MOSSY_MANGROVE_PLANKS);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_CHERRY_VERTICAL_SLAB, CRIMSON_MOSSY_CHERRY_PLANKS);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_PALE_OAK_VERTICAL_SLAB, CRIMSON_MOSSY_CHERRY_PLANKS);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_BAMBOO_VERTICAL_SLAB, CRIMSON_MOSSY_BAMBOO_PLANKS);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_BAMBOO_MOSAIC_VERTICAL_SLAB, CRIMSON_MOSSY_BAMBOO_MOSAIC);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_CRIMSON_VERTICAL_SLAB, CRIMSON_MOSSY_CRIMSON_PLANKS);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_WARPED_VERTICAL_SLAB, CRIMSON_MOSSY_WARPED_PLANKS);
+                createVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_STONE_VERTICAL_SLAB, CRIMSON_MOSSY_STONE);
+                createVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_COBBLED_DEEPSLATE_VERTICAL_SLAB, CRIMSON_MOSSY_COBBLED_DEEPSLATE);
+                createVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_DEEPSLATE_BRICK_VERTICAL_SLAB, CRIMSON_MOSSY_DEEPSLATE_BRICKS);
+                createVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_DEEPSLATE_TILE_VERTICAL_SLAB, CRIMSON_MOSSY_DEEPSLATE_TILES);
+                createVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_TUFF_VERTICAL_SLAB, CRIMSON_MOSSY_TUFF);
+                createVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_TUFF_BRICK_VERTICAL_SLAB, CRIMSON_MOSSY_TUFF_BRICKS);
+                createVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_BRICK_VERTICAL_SLAB, CRIMSON_MOSSY_BRICKS);
+
+                this.stonecutterResultFromBase(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_STONE_VERTICAL_SLAB, CRIMSON_MOSSY_STONE, 2);
+                this.stonecutterResultFromBase(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_COBBLED_DEEPSLATE_VERTICAL_SLAB, CRIMSON_MOSSY_COBBLED_DEEPSLATE, 2);
+                this.stonecutterResultFromBase(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_DEEPSLATE_BRICK_VERTICAL_SLAB, CRIMSON_MOSSY_DEEPSLATE_BRICKS, 2);
+                this.stonecutterResultFromBase(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_DEEPSLATE_TILE_VERTICAL_SLAB, CRIMSON_MOSSY_DEEPSLATE_TILES, 2);
+                this.stonecutterResultFromBase(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_TUFF_VERTICAL_SLAB, CRIMSON_MOSSY_TUFF, 2);
+                this.stonecutterResultFromBase(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_TUFF_BRICK_VERTICAL_SLAB, CRIMSON_MOSSY_TUFF_BRICKS, 2);
+                this.stonecutterResultFromBase(RecipeCategory.BUILDING_BLOCKS, ModItems.CRIMSON_MOSSY_BRICK_VERTICAL_SLAB, CRIMSON_MOSSY_BRICKS, 2);
+                
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_OAK_VERTICAL_SLAB, WARPED_MOSSY_OAK_PLANKS);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_SPRUCE_VERTICAL_SLAB, WARPED_MOSSY_SPRUCE_PLANKS);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_BIRCH_VERTICAL_SLAB, WARPED_MOSSY_BIRCH_PLANKS);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_JUNGLE_VERTICAL_SLAB, WARPED_MOSSY_JUNGLE_PLANKS);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_ACACIA_VERTICAL_SLAB, WARPED_MOSSY_ACACIA_PLANKS);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_DARK_OAK_VERTICAL_SLAB, WARPED_MOSSY_DARK_OAK_PLANKS);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_MANGROVE_VERTICAL_SLAB, WARPED_MOSSY_MANGROVE_PLANKS);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_CHERRY_VERTICAL_SLAB, WARPED_MOSSY_CHERRY_PLANKS);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_PALE_OAK_VERTICAL_SLAB, WARPED_MOSSY_CHERRY_PLANKS);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_BAMBOO_VERTICAL_SLAB, WARPED_MOSSY_BAMBOO_PLANKS);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_BAMBOO_MOSAIC_VERTICAL_SLAB, WARPED_MOSSY_BAMBOO_MOSAIC);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_CRIMSON_VERTICAL_SLAB, WARPED_MOSSY_CRIMSON_PLANKS);
+                createWoodenVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_WARPED_VERTICAL_SLAB, WARPED_MOSSY_WARPED_PLANKS);
+                createVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_STONE_VERTICAL_SLAB, WARPED_MOSSY_STONE);
+                createVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_COBBLED_DEEPSLATE_VERTICAL_SLAB, WARPED_MOSSY_COBBLED_DEEPSLATE);
+                createVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_DEEPSLATE_BRICK_VERTICAL_SLAB, WARPED_MOSSY_DEEPSLATE_BRICKS);
+                createVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_DEEPSLATE_TILE_VERTICAL_SLAB, WARPED_MOSSY_DEEPSLATE_TILES);
+                createVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_TUFF_VERTICAL_SLAB, WARPED_MOSSY_TUFF);
+                createVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_TUFF_BRICK_VERTICAL_SLAB, WARPED_MOSSY_TUFF_BRICKS);
+                createVerticalSlabRecipe(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_BRICK_VERTICAL_SLAB, WARPED_MOSSY_BRICKS);
+
+                this.stonecutterResultFromBase(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_STONE_VERTICAL_SLAB, WARPED_MOSSY_STONE, 2);
+                this.stonecutterResultFromBase(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_COBBLED_DEEPSLATE_VERTICAL_SLAB, WARPED_MOSSY_COBBLED_DEEPSLATE, 2);
+                this.stonecutterResultFromBase(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_DEEPSLATE_BRICK_VERTICAL_SLAB, WARPED_MOSSY_DEEPSLATE_BRICKS, 2);
+                this.stonecutterResultFromBase(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_DEEPSLATE_TILE_VERTICAL_SLAB, WARPED_MOSSY_DEEPSLATE_TILES, 2);
+                this.stonecutterResultFromBase(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_TUFF_VERTICAL_SLAB, WARPED_MOSSY_TUFF, 2);
+                this.stonecutterResultFromBase(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_TUFF_BRICK_VERTICAL_SLAB, WARPED_MOSSY_TUFF_BRICKS, 2);
+                this.stonecutterResultFromBase(RecipeCategory.BUILDING_BLOCKS, ModItems.WARPED_MOSSY_BRICK_VERTICAL_SLAB, WARPED_MOSSY_BRICKS, 2);
             }
 
             public void createWoodenVerticalSlabRecipe(RecipeCategory category, ItemLike output, ItemLike input) {
@@ -92,7 +151,7 @@ public class ModRecipeProvider extends FabricRecipeProvider {
                         .pattern("#")
                         .pattern("#")
                         .pattern("#")
-                        .unlockedBy("has_wood", has(ModItemTags.MOSSY_PLANKS))
+                        .unlockedBy("has_wood", has(MOSSY_PLANKS))
                         .group("wooden_mossy_vertical_slabs")
                         .save(this.output);
             }